feat: rework in a more efficient way commands and handlers

+++ src/main.py
@@ -5,8 +5,6 @@ import json
 import logging
 import functools
 import asyncio
-import plotly.graph_objs as go
-import plotly.io as pio
 import matplotlib.pyplot as plt
 from datetime import datetime
 from typing import Tuple, List
@@ -21,46 +19,15 @@ PNG_FILE_NAME = 'plot.png'
 BUDDY_FILE_NAME = 'Buddy_christ.jpg'
 PAT_FILE_NAME = 'patrick.gif'
 
-# String constants
-HI_CMD_TXT = 'hi'
-HI_CMD_DESC_TXT = 'Say hi to Robbi'
-NODE_CMD_TXT = 'node'
-NODE_CMD_DESC_TXT = 'Get your node results'
-BTC_CMD_TXT = 'btc'
-BTC_CMD_DESC_TXT = 'Get BTC current price'
-MAS_CMD_TXT = 'mas'
-MAS_CMD_DESC_TXT = 'Get MAS current price'
-FLUSH_CMD_TXT = 'flush'
-FLUSH_CMD_DESC_TXT = 'Flush local logs'
-NODE_IS_DOWN = 'Node is down'
-
 COMMANDS_LIST = [
-    {
-        'id': 0,
-        'cmd_txt': HI_CMD_TXT,
-        'cmd_desc': HI_CMD_DESC_TXT
-    },
-    {
-        'id': 1,
-        'cmd_txt': NODE_CMD_TXT,
-        'cmd_desc': NODE_CMD_DESC_TXT
-    },
-    {
-        'id': 2,
-        'cmd_txt': BTC_CMD_TXT,
-        'cmd_desc': BTC_CMD_DESC_TXT
-    },
-    {
-        'id': 3,
-        'cmd_txt': MAS_CMD_TXT,
-        'cmd_desc': MAS_CMD_DESC_TXT
-    },
-    {
-        'id': 4,
-        'cmd_txt': FLUSH_CMD_TXT,
-        'cmd_desc': FLUSH_CMD_DESC_TXT
-    }
+    {'id': 0, 'cmd_txt': 'hi', 'cmd_desc': 'Say hi to Robbi'},
+    {'id': 1, 'cmd_txt': 'node', 'cmd_desc': 'Get your node results'},
+    {'id': 2, 'cmd_txt': 'btc', 'cmd_desc': 'Get BTC current price'},
+    {'id': 3, 'cmd_txt': 'mas', 'cmd_desc': 'Get MAS current price'},
+    {'id': 4, 'cmd_txt': 'flush', 'cmd_desc': 'Flush local logs'},
 ]
+
+NODE_IS_DOWN = 'Node is down'
 
 # Configure logging module
 logging.basicConfig(
@@ -101,7 +68,7 @@ def extract_address_data(json_data: dict) -> Tuple[str, int, List[int], List[int
         return final_balance, final_roll_count, cycles, ok_counts, nok_counts, active_rolls
     return "", 0, [], [], [], []
 
-def create_png_pot(cycles: int, nok_counts: int, ok_counts: int) -> str:
+def create_png_plot(cycles: int, nok_counts: int, ok_counts: int) -> str:
         """
         Creates a line plot with markers for OK and NOK counts over multiple cycles,
         and saves the plot as a PNG image.
@@ -124,19 +91,15 @@ def create_png_pot(cycles: int, nok_counts: int, ok_counts: int) -> str:
         plt.close()
         return PNG_FILE_NAME
 
-async def hello(update: Update, context: CallbackContext) -> None:
+async def hi(update: Update, context: CallbackContext) -> None:
     user_id = str(update.effective_user.id)
-    logging.info(f'User {user_id} used the /hello command.')
+    logging.info(f'User {user_id} used the /hi command.')
 
-    if user_id in allowed_user_ids:
-        photo_path = 'media/' + BUDDY_FILE_NAME
-        await update.message.reply_text('Hey dude!')
-        await update.message.reply_photo(photo=photo_path)
-    else:
-        photo_path = 'media/' + PAT_FILE_NAME
-        await update.message.reply_photo(photo=photo_path)
+    photo_path = 'media/' + (BUDDY_FILE_NAME if user_id in allowed_user_ids else PAT_FILE_NAME)
+    await update.message.reply_text('Hey dude!' if user_id in allowed_user_ids else '')
+    await update.message.reply_photo(photo=photo_path)
 
-async def massa_node(update: Update, context: CallbackContext) -> None:
+async def node(update: Update, context: CallbackContext) -> None:
     user_id = str(update.effective_user.id)
     logging.info(f'User {user_id} used the /node command.')
 
@@ -145,20 +108,19 @@ async def massa_node(update: Update, context: CallbackContext) -> None:
         json_data = get_addresses(logging, massa_node_address)
 
         # Extract useful data using the function
-        final_balance, final_roll_count, cycles, ok_counts, nok_counts, active_rolls = extract_address_data(json_data)
+        data = extract_address_data(json_data)
         formatted_string = (
-            f"Final Balance: {final_balance}\n"
-            f"Final Roll Count: {final_roll_count}\n"
-            f"OK Counts: {ok_counts}\n"
-            f"NOK Counts: {nok_counts}\n"
-            f"Active Rolls: {active_rolls}"
+            f"Final Balance: {data[0]}\n"
+            f"Final Roll Count: {data[1]}\n"
+            f"OK Counts: {data[3]}\n"
+            f"NOK Counts: {data[4]}\n"
+            f"Active Rolls: {data[5]}"
         )
         print(formatted_string)
         await update.message.reply_text('Node status: ' + formatted_string)
 
         # Create graph from data and save to PNG_FILE_NAME
-        image_path = create_png_pot(cycles, nok_counts, ok_counts)
-
+        image_path = create_png_plot(data[2], data[4], data[3])
         # Check if the image file was created successfully
         if os.path.exists(image_path):
             try:
@@ -171,27 +133,25 @@ async def massa_node(update: Update, context: CallbackContext) -> None:
         else:
             logging.error("Image file was not created successfully.")
 
-async def remove_logs(update: Update, context: CallbackContext) -> None:
+async def flush(update: Update, context: CallbackContext) -> None:
     user_id = str(update.effective_user.id)
-    logging.info(f'User {user_id} used the /flush /clean command.')
+    logging.info(f'User {user_id} used the /flush command.')
 
-    if user_id in allowed_user_ids:
-        if os.path.exists(LOG_FILE_NAME):
-            try:
-                # Clear the log file at runtime
-                with open(LOG_FILE_NAME, 'w'):
-                    pass  # This will clear the file content
-                message = f"{LOG_FILE_NAME} has been cleared."
-                print(message)
-                await update.message.reply_text(message)
-            except IOError as e:
-                logging.error(f"Error clearing the log file: {e}")
-                await update.message.reply_text("An error occurred while clearing the log file.")
-        else:
-            logging.warning(f"Log file {LOG_FILE_NAME} does not exist.")
-            await update.message.reply_text(f"Log file {LOG_FILE_NAME} does not exist.")
+    if user_id in allowed_user_ids and os.path.exists(LOG_FILE_NAME):
+        try:
+            with open(LOG_FILE_NAME, 'w'):
+                pass
+            message = f"{LOG_FILE_NAME} has been cleared."
+            print(message)
+            await update.message.reply_text(message)
+        except IOError as e:
+            logging.error(f"Error clearing the log file: {e}")
+            await update.message.reply_text("An error occurred while clearing the log file.")
+    elif not os.path.exists(LOG_FILE_NAME):
+        logging.warning(f"Log file {LOG_FILE_NAME} does not exist.")
+        await update.message.reply_text(f"Log file {LOG_FILE_NAME} does not exist.")
 
-async def bitcoin(update: Update, context: CallbackContext) -> None:
+async def btc(update: Update, context: CallbackContext) -> None:
     user_id = str(update.effective_user.id)
     logging.info(f'User {user_id} used the /btc command.')
 
@@ -229,19 +189,10 @@ async def mas(update: Update, context: CallbackContext) -> None:
         print(formatted_string)
         await update.message.reply_text(formatted_string)
 
-HANDLERS = [
-    (COMMANDS_LIST[0]['cmd_txt'], hello),
-    (COMMANDS_LIST[1]['cmd_txt'], massa_node),
-    (COMMANDS_LIST[2]['cmd_txt'], bitcoin),
-    (COMMANDS_LIST[3]['cmd_txt'], mas),
-    (COMMANDS_LIST[4]['cmd_txt'], remove_logs)
-]
+HANDLERS = [(cmd['cmd_txt'], globals()[cmd['cmd_txt']]) for cmd in COMMANDS_LIST]
 
 async def post_init(application: Application) -> None:
-    commands = [
-        BotCommand(command=cmd['cmd_txt'], description=cmd['cmd_desc'])
-        for cmd in COMMANDS_LIST
-    ]
+    commands = [BotCommand(command=cmd['cmd_txt'], description=cmd['cmd_desc']) for cmd in COMMANDS_LIST]
     await application.bot.set_my_commands(commands)
 
 def run_async_func(application: Application) -> None:
@@ -259,7 +210,7 @@ def run_async_func(application: Application) -> None:
 
     scheduler.start()
 
-def run_coroutine_in_loop(coroutine, loop):
+def run_coroutine_in_loop(coroutine, loop) -> None:
     asyncio.set_event_loop(loop)
     loop.run_until_complete(coroutine)
 
@@ -267,12 +218,10 @@ async def periodic_node_ping(application: Application) -> None:
     logging.info(f'Node ping')
     json_data = get_addresses(logging, massa_node_address)
     # Extract useful data using the function
-    final_balance, final_roll_count, cycles, ok_counts, nok_counts, active_rolls = extract_address_data(json_data)
-    logging.info((any(element != 0 for element in nok_counts) != 0))
-    logging.info(active_rolls != prev_active_rolls)
-    if((any(element != 0 for element in nok_counts) != 0) or (active_rolls != prev_active_rolls)):
+    data = extract_address_data(json_data)
+    if any(data[4]) or data[5] != prev_active_rolls:
         for user_id in allowed_user_ids:
-            run_async_func(application) # Restart the scheduler
+            run_async_func(application)
             await application.bot.send_message(chat_id=user_id, text=NODE_IS_DOWN)
 
 
@@ -292,11 +241,8 @@ def main():
             allowed_user_ids = {str(tmp_json.get('user_white_list', {}).get('admin'))}
             massa_node_address = tmp_json.get('massa_node_address')
             ninja_key = tmp_json.get('ninja_api_key')
-    except FileNotFoundError:
-        logging.error("The file 'topology.json' was not found.")
-        return
-    except json.JSONDecodeError:
-        logging.error("Error decoding the JSON file.")
+    except (FileNotFoundError, json.JSONDecodeError) as e:
+        logging.error(f"Error loading topology.json: {e}")
         return
 
     disable_prints()
@@ -304,8 +250,7 @@ def main():
 
     # Get node info at bot startup
     json_data = get_addresses(logging, massa_node_address)
-    final_balance, final_roll_count, cycles, ok_counts, nok_counts, active_rolls = extract_address_data(json_data)
-    prev_active_rolls = list(active_rolls)
+    prev_active_rolls = list(extract_address_data(json_data)[5])
 
     # Use of ApplicationBuilder to create app
     application = Application.builder().token(bot_token).post_init(post_init).build()
